requirements.txt
pillow>=8.3
python-telegram-bot==13.7
opencv-python-headless
redis-py
